The Seller is added

+++ src/main/java/com/spring/jwt/repository/BuyerRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BuyerRepository extends JpaRepository<Buyer, Integer> {
+public interface BuyerRepository extends JpaRepository<Buyer, Long> {
 
      // Spring Data JPA will automatically implement this method
-     Buyer findByUserId(Integer userId);  //  (notice user.id)
+     Buyer findByUserId(Long userId);
 
 }
